# openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Splite API",
    "version": "2.0.0-contract",
    "description": "Bill splitting for Venezuelan restaurants.\n\nSettlement is always VES in céntimos. USD and EUR are display references, never payment currencies.\n\n## Wire format conventions\n\nEvery endpoint follows these rules. A frontend that trusts them never needs to guess.\n\n| Kind | JSON type | Format | Example |\n|------|-----------|--------|---------|\n| **Money** (minor units) | `string` | Digit string — no leading zeros except `\"0\"` | `\"9007199254740993\"` |\n| **FX rates** | `string` | Decimal, padded to 8 fractional digits | `\"757.54060000\"` |\n| **IDs** | `string` | UUID v4 | `\"d290f1ee-6c54-4b01-90e6-d701748f0851\"` |\n| **Timestamps** | `string` | ISO 8601 `date-time` | `\"2025-03-05T16:30:00.000Z\"` |\n| **Value dates** | `string` | ISO 8601 `date` | `\"2025-03-06\"` |\n\nAmounts are strings so values beyond 2^53 survive JSON (JavaScript `Number` loses\nprecision past `Number.MAX_SAFE_INTEGER`). Rates are strings so every endpoint returns\nthe same representation — no `757.5406` from one route and `\"757.54060000\"` from another.\n\nEvery query is scoped to the caller's restaurant. A resource belonging to another tenant\nis reported as 404 rather than 403, so an endpoint never confirms that it exists.",
    "x-error-details": {
      "VALIDATION_FAILED": {
        "fields": "string[] -- one entry per failed field"
      },
      "OPEN_BILL_EXISTS": {
...
  "servers": [
...
  "security": [
    {
      "staffAuth": []
    }
  ],
  "components": {
    "securitySchemes": {
      "staffAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "description": "Short-lived staff access token from /api/v1/auth/login."
      },
      "guestAuth": {
        "type": "http",
        "scheme": "bearer",
        "description": "Guest session token. Send the token as a bearer credential *and* the session id in the `X-Guest-Session` header; both are required. Obtained from POST /api/v1/guest/sessions by presenting a signed table QR."
      }
    },
...
    "schemas": {
      "Error": {
        "type": "object",
        "required": [
          "error"
        ],
        "properties": {
          "error": {
            "type": "object",
            "required": [
              "code",
              "message",
              "details",
              "requestId"
            ],
            "properties": {
              "code": {
                "type": "string",
                "enum": [
                  "VALIDATION_FAILED",
                  "BILL_ID_MISMATCH",
                  "INVALID_AMOUNT",
                  "INVALID_MONETARY_VALUE",
                  "WEBHOOK_BODY_UNVERIFIABLE",
                  "SPLIT_PARTICIPANTS_INVALID",
                  "SPLIT_CLAIMS_INCOMPLETE",
                  "SPLIT_CLAIM_UNKNOWN",
                  "SPLIT_AMOUNT_MISMATCH",
                  "AUTH_TOKEN_MISSING",
                  "AUTH_TOKEN_INVALID",
                  "INVALID_CREDENTIALS",
                  "GUEST_SESSION_MISSING",
                  "GUEST_SESSION_INVALID",
                  "QR_INVALID",
                  "WEBHOOK_SIGNATURE_MISSING",
                  "WEBHOOK_SIGNATURE_INVALID",
                  "WEBHOOK_TIMESTAMP_OUT_OF_TOLERANCE",
                  "FORBIDDEN_ROLE",
                  "CROSS_TENANT_DENIED",
                  "CORS_ORIGIN_NOT_ALLOWED",
                  "NOT_FOUND",
                  "BILL_NOT_FOUND",
                  "BILL_ITEM_NOT_FOUND",
                  "TABLE_NOT_FOUND",
                  "PRODUCT_NOT_FOUND",
                  "RESTAURANT_NOT_FOUND",
                  "OPEN_BILL_NOT_FOUND",
                  "OPEN_BILL_EXISTS",
                  "BILL_NOT_OPEN",
                  "BILL_NOT_ITEMISED",
                  "TOTAL_BELOW_AMOUNT_PAID",
                  "PRODUCT_INACTIVE",
                  "SPLIT_NOT_ITEMISED",
                  "SPLIT_NOTHING_OUTSTANDING",
                  "BILL_HAS_PAYMENTS",
                  "PAYMENT_EXCEEDS_BALANCE",
                  "PAYMENT_STATE_INVALID",
                  "IDEMPOTENCY_KEY_REUSED",
                  "IDEMPOTENCY_IN_FLIGHT",
                  "TABLE_NAME_TAKEN",
                  "PRODUCT_NAME_TAKEN",
                  "MENU_CURRENCY_MISMATCH",
                  "WEBHOOK_ALREADY_PROCESSED",
                  "RATE_LIMITED",
                  "RATE_LIMITER_UNAVAILABLE",
                  "FX_UNAVAILABLE",
                  "WEBHOOK_REPLAY_PROTECTION_UNAVAILABLE",
                  "SHUTTING_DOWN",
                  "INTERNAL_ERROR"
                ],
                "description": "Stable identifier for what went wrong. Branch on this, never on `message`. A code always carries the same HTTP status."
              },
              "message": {
                "type": "string",
                "description": "Human-readable and subject to change without notice. Never parse it. 5xx messages are always the literal string \"Internal Server Error\"."
              },
              "details": {
                "type": "object",
                "additionalProperties": true,
                "description": "Structured context for this code, always present and possibly empty. See x-error-details for what each code carries."
              },
              "requestId": {
                "type": "string",
                "description": "Correlates with the server log line for this failure. Quote it in bug reports."
              }
            }
          }
        }
      },
      "ValidationError": {
        "allOf": [
          {
            "$ref": "#/components/schemas/Error"
          },
          {
            "type": "object",
            "description": "code is always VALIDATION_FAILED; details.fields carries one entry per failed field.",
            "properties": {
              "error": {
                "type": "object",
                "properties": {
                  "details": {
                    "type": "object",
                    "properties": {
                      "fields": {
                        "type": "array",
                        "items": {
                          "type": "string"
                        }
                      }
                    }
                  }
                }
              }
            }
          }
        ]
      },
...
      "LoginRequest": {
        "type": "object",
        "required": [
          "email",
          "password"
        ],
        "properties": {
          "email": {
            "type": "string",
            "format": "email",
            "maxLength": 254
          },
          "password": {
            "type": "string",
            "minLength": 1,
            "maxLength": 128
          }
        }
      },
      "RefreshRequest": {
        "type": "object",
        "required": [
          "refreshToken"
        ],
        "properties": {
          "refreshToken": {
            "type": "string",
            "minLength": 20,
            "maxLength": 4096
          }
        }
      },
      "Session": {
        "type": "object",
        "properties": {
          "accessToken": {
            "type": "string"
          },
          "refreshToken": {
            "type": "string"
          },
          "expiresIn": {
            "type": "string"
          },
          "user": {
            "$ref": "#/components/schemas/SessionUser"
          }
        }
      },
      "SessionUser": {
        "type": "object",
        "description": "The authenticated staff member. Returned by login, refresh and /auth/me alike, so a client stores one type.",
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "email": {
            "type": "string"
          },
          "role": {
            "type": "string",
            "enum": [
              "OWNER",
              "MANAGER",
              "CASHIER",
              "WAITER"
            ]
          },
          "restaurantId": {
            "type": "string",
            "format": "uuid"
          }
        }
      },
...
    },
    "responses": {
      "BadRequest": {
        "description": "Validation failed, or the body contradicts the path.",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/ValidationError"
            }
          }
        }
      },
      "Unauthorized": {
        "description": "Missing, malformed or expired credentials.",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/Error"
            }
          }
        }
      },
...
      "TooManyRequests": {
        "description": "Rate limit exceeded, or the limiter is unavailable on a fail-closed surface.",
        "headers": {
          "Retry-After": {
            "schema": {
              "type": "integer"
            },
            "description": "Seconds until the window resets."
          }
        },
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/Error"
            }
          }
        }
      },
      "ServerError": {
        "description": "Unexpected failure. The message is never echoed; correlate using requestId.",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/Error"
            }
          }
        }
      },
      "ServiceUnavailable": {
        "description": "A dependency is unavailable.",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/Error"
            }
          }
        }
      }
    }
  },
  "paths": {
...
    "/api/v1/auth/login": {
      "post": {
        "tags": [
          "Auth"
        ],
        "summary": "Exchange credentials for a session",
        "description": "Rate limited to 10/minute per IP, and fails closed in production: if Redis is unavailable this returns 503 rather than waving brute-force attempts through. Unknown and known emails take the same time.",
        "security": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/LoginRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Session issued.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Session"
                }
              }
            }
          },
          "400": {
            "$ref": "#/components/responses/BadRequest"
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          },
          "429": {
            "$ref": "#/components/responses/TooManyRequests"
          },
          "500": {
            "$ref": "#/components/responses/ServerError"
          },
          "503": {
            "$ref": "#/components/responses/ServiceUnavailable"
          }
        }
      }
    },
    "/api/v1/auth/me": {
      "get": {
        "tags": [
          "Auth"
        ],
        "summary": "The current user",
        "operationId": "getCurrentUser",
        "description": "What a client calls when restoring a session on boot. Any authenticated staff role.\n\n**Use this rather than /auth/refresh to identify the caller.** Refresh *rotates*: two tabs\nstarting at once both present the same stored token, one claims it, and the other is\ntreated as theft and revokes every session for that user. Calling refresh merely to ask\n\"who am I\" turns a second browser tab into a logout.\n\nRead from the database rather than the token, so a deactivated account stops working\ninside the access token's fifteen minutes rather than at the end of them.",
        "security": [
          {
            "staffAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "The caller. Identical in shape to `user` in a login or refresh response.",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "user": {
                      "$ref": "#/components/schemas/SessionUser"
                    }
                  }
                }
              }
            }
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          },
          "429": {
            "$ref": "#/components/responses/TooManyRequests"
          },
          "500": {
            "$ref": "#/components/responses/ServerError"
          }
        }
      }
    },
    "/api/v1/auth/refresh": {
      "post": {
        "tags": [
          "Auth"
        ],
        "summary": "Rotate a refresh token",
        "description": "Rotation is atomic. Presenting an already-revoked token is treated as theft and revokes every session for that user.",
        "security": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RefreshRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "New session issued.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Session"
                }
              }
            }
          },
          "400": {
            "$ref": "#/components/responses/BadRequest"
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          },
          "429": {
            "$ref": "#/components/responses/TooManyRequests"
          },
          "500": {
            "$ref": "#/components/responses/ServerError"
          }
        }
      }
    },
    "/api/v1/auth/logout": {
      "post": {
        "tags": [
          "Auth"
        ],
        "summary": "Revoke a refresh session",
        "description": "Always 204, so it never reveals whether the token was valid.",
        "security": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RefreshRequest"
              }
            }
          }
        },
        "responses": {
          "204": {
            "description": "Revoked, or the token was already invalid."
          },
          "400": {
            "$ref": "#/components/responses/BadRequest"
          },
          "429": {
            "$ref": "#/components/responses/TooManyRequests"
          },
          "500": {
            "$ref": "#/components/responses/ServerError"
          }
        }
      }
    },
...
  }
}